Remove a some of spaghetti code.
